pyproject.toml
[tool.poetry]
name = "app-review"
version = "0.1.0"
description = ""
authors = ["smpfy <arsmpfy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
app-store-scraper = "^0.3.5"
emoji = "^2.9.0"
numpy = "^1.26.3"
pyldavis = "^3.4.1"
hydra-core = "^1.3.2"
google-play-scraper = "^1.2.4"
psycopg2-binary = "^2.9.9"
pyspark = "3.3.4"
spark-nlp = "^5.2.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[tool.poetry.group.webapp.dependencies]
fastapi = "^0.108.0"
uvicorn = "^0.25.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
